Binomial implemetation

+++ Binomial.java
@@ -7,11 +7,12 @@ public class Binomial {
     	// System.out.println(binomial1(Integer.parseInt(args[0]), Integer.parseInt(args[1])));
 
 		// Testing the optimized binomial implementation:
-		 System.out.println(binomial(Integer.parseInt(args[0]), Integer.parseInt(args[1])));
+		  System.out.println(binomial(Integer.parseInt(args[0]), Integer.parseInt(args[1])));
+		//System.out.println(binomial(5, 3));
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) { 
+	public static long binomial1(int n, int k) { 
 		if (k > n) return 0;
 		if (k == 0 || n == 0) return 1;
 		return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
@@ -19,9 +20,9 @@ public class Binomial {
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
+	public static long binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, 
-		//// and then initializes all its elements to -1.
+		// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
 		int rows = n + 1;
 		int cols = k + 1;
